Add a BlacklightOverrides concern for SearchAcross behavior.
