+++ .rubocop_todo.yml
@@ -1,32 +1,21 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2020-02-25 09:28:53 -0800 using RuboCop version 0.80.0.
+# on 2020-02-27 21:19:26 -0800 using RuboCop version 0.80.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# Offense count: 2
+# Offense count: 1
 # Cop supports --auto-correct.
-# Configuration parameters: EnforcedStyle, IndentOneStep, IndentationWidth.
-# SupportedStyles: case, end
-Layout/CaseIndentation:
-  Exclude:
-    - 'app/helpers/search_across_helper.rb'
-
-# Offense count: 6
-# Cop supports --auto-correct.
-# Configuration parameters: AllowMultipleStyles, EnforcedHashRocketStyle, EnforcedColonStyle, EnforcedLastArgumentHashStyle.
-# SupportedHashRocketStyles: key, separator, table
-# SupportedColonStyles: key, separator, table
-# SupportedLastArgumentHashStyles: always_inspect, always_ignore, ignore_implicit, ignore_explicit
-Layout/HashAlignment:
-  Exclude:
-    - 'app/helpers/search_across_helper.rb'
+# Configuration parameters: AutoCorrect, AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
+# URISchemes: http, https
+Layout/LineLength:
+  Max: 122
 
 # Offense count: 12
 Metrics/AbcSize:
-  Max: 39
+  Max: 42
 
 # Offense count: 1
 Metrics/CyclomaticComplexity:
@@ -67,7 +56,7 @@ RSpec/ExpectInHook:
     - 'spec/views/catalog/_exhibits_document_header_default.html.erb_spec.rb'
     - 'spec/views/viewers/edit.html.erb_spec.rb'
 
-# Offense count: 112
+# Offense count: 113
 # Configuration parameters: AggregateFailuresByDefault.
 RSpec/MultipleExpectations:
   Max: 6
@@ -81,7 +70,7 @@ RSpec/SubjectStub:
   Exclude:
     - 'spec/models/dor_harvester_spec.rb'
 
-# Offense count: 4
+# Offense count: 3
 # Configuration parameters: Include.
 # Include: app/helpers/**/*.rb
 Rails/HelperInstanceVariable:
@@ -91,7 +80,7 @@ Rails/HelperInstanceVariable:
 # Offense count: 3
 Rails/OutputSafety:
   Exclude:
-    - 'app/helpers/search_across_helper.rb'
+    - 'app/controllers/concerns/search_across_blacklight_overrides.rb'
 
 # Offense count: 1
 # Cop supports --auto-correct.
