+++ spec/helpers/search_across_helper_spec.rb
@@ -137,50 +137,4 @@ RSpec.describe SearchAcrossHelper, type: :helper do
       expect(helper.exhibit_metadata.keys).to match_array(%w(abc xyz))
     end
   end
-
-  describe 'Blacklight overrides' do
-    describe '#show_pagination?' do
-      it 'is suppressed for grouped responses' do
-        allow(helper).to receive(:render_grouped_response?).and_return(true)
-        expect(helper.show_pagination?).to eq false
-      end
-    end
-
-    describe '#document_index_path_templates' do
-      it 'injects custom exhibits document partials' do
-        allow(helper).to receive(:render_grouped_response?).and_return(true)
-        expect(helper.document_index_path_templates).to eq ['exhibit_%<index_view_type>s']
-      end
-    end
-
-    describe '#render_grouped_document_index' do
-      let(:response) do
-        Blacklight::Solr::Response.new({
-                                         facet_counts: {
-                                           facet_fields: {
-                                             "#{SolrDocument.exhibit_slug_field}": {
-                                               a: 1,
-                                               b: 3,
-                                               c: 15
-                                             }
-                                           }
-                                         }
-                                       }, nil)
-      end
-
-      let(:exhibits) do
-        [
-          create(:exhibit, slug: 'a'),
-          create(:exhibit, slug: 'b'),
-          create(:exhibit, slug: 'c')
-        ]
-      end
-
-      it 'replaces solr results with exhibits from the database' do
-        allow(helper).to receive(:render_document_index).with(exhibits).and_return('x')
-        helper.render_grouped_document_index(response)
-        expect(helper).to have_received(:render_document_index).with(exhibits)
-      end
-    end
-  end
 end
